added an option of plotting 3 values and improved graphing speed
For now it can only receive three values separated by spaces, but the
there is no option to select which value will be plotted, it just plots
them in order. I have removed the time.sleep command in the ComMonitor
and it improved the graphing speed drastically. I have also noticed that
if the time.sleep delay was longer then the delay in the Arduino loop,
it had a huge delay while plotting, since it plotted values from an
older time.
Next task: 1. add data plotting selector 2. add menu to select COMport,
baudrate,...